Add a fruit salad recipe page
Add a web page for a fruit salad recipe to the project. The page
includes an image of a fruit salad, a decription of the recipe, an
ingredient list, and the steps to make the dish. Also add a link to the
fruit salad recipe page to the home page.